requirements.txt
python-dotenv
openai>=1.30.0
chromadb>=0.5.0
pypdf
streamlit
pyttsx3
